# Vectorize `unique`