--- a/benchmarks/src/unique.cpp
+++ b/benchmarks/src/unique.cpp
@@ -5,6 +5,7 @@
 #include <benchmark/benchmark.h>
 #include <cstdint>
 #include <random>
+#include <type_traits>
 #include <vector>
 
 #include "skewed_allocator.hpp"