--- a/tests/std/tests/VSO_0000000_vector_algorithms/test.cpp
+++ b/tests/std/tests/VSO_0000000_vector_algorithms/test.cpp
@@ -13,7 +13,6 @@
 #include <functional>
 #include <limits>
 #include <list>
-#include <memory>
 #include <numeric>
 #include <random>
 #include <stdexcept>