@StephanTLavavej 2025-03-03
Thanks! (And sorry for taking a while to get to this. Partial review: still need to finish reviewing `test.cpp` and `vector_algorithms.cpp`.)

I pushed a conflict-free merge with `main` followed by tiny nitpicks, but then noticed a bug in the test.