@StephanTLavavej 2025-03-04
The speedups are less significant on my 5950X, but good across the board with no regressions:

Benchmark                            |  Before |   After | Speedup
-------------------------------------|--------:|--------:|--------
`u<alg_type::std_fn, std::uint8_t>`  | 1681 ns |  494 ns | 3.40
`u<alg_type::std_fn, std::uint16_t>` | 1566 ns |  503 ns | 3.11
`u<alg_type::std_fn, std::uint32_t>` | 1319 ns |  657 ns | 2.01
`u<alg_type::std_fn, std::uint64_t>` | 1466 ns | 1138 ns | 1.29
`u<alg_type::rng, std::uint8_t>`     | 1261 ns |  497 ns | 2.54
`u<alg_type::rng, std::uint16_t>`    | 1691 ns |  584 ns | 2.90
`u<alg_type::rng, std::uint32_t>`    | 1765 ns |  601 ns | 2.94
`u<alg_type::rng, std::uint64_t>`    | 1403 ns | 1139 ns | 1.23